delete test pipeline
